requirements.txt
Django==3.2.8

django-cors-headers==3.10.0

mysqlclient==2.0.3

PyJWT==2.3.0

bcrypt==3.2.0